requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.9
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.3
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
sqlparse==0.4.4
typing-extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
